requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
playwright
httpx
pytest
scrapy
